neptune2/requirements.txt
google-adk
toolbox-core
google-genai
